Bug 1627012 - Ensure NotifyLayersUpdated doesn't get called with bogus scrollable rects. r=botond
SetScrollableFrameMetrics uses a bogus scrollable rect (-1, -1, -1, -1) if none
is provided. With the next patch, we start using this bogus rect for clamping
the scroll offset on some calls to NotifyLayersUpdated, and causes the gtest
to fail. If we provide some sane rect it all works fine.

Depends on D76810

Differential Revision: https://phabricator.services.mozilla.com/D76811

UltraBlame original commit: 76db5106df729a063326eb9d810426a6a97037fe