Bug 1637994 - Add a transformer for extracting times/durations from logcat logs. r=acreskey,mozperftest-reviewers,tarek
This patch adds a transformer that is capable of extracting times/entries or durations for logcat logs. Here, we also implement the ability for users to pass options to the transformers so that they can be more generic (through transformer-options). The logcat transformer makes use of this, and also shows how a custom function can be used within transformers. A new test file is added for the transformer, as well as a test in test_perfherder.py so we can test the full pipeline.

Differential Revision: https://phabricator.services.mozilla.com/D77094

UltraBlame original commit: 2d5f01c88b7e4b56319feea1261d546a0129a646