Bug 1641711 - Make javadoc output parsing more resilient. r=snorp
As it turns out, javadoc output is not stable across platforms / OSs / versions
(not a big suprise).

E.g. in try this is an example line:

```
/builds/worker/checkouts/gecko/mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoRuntimeSettings.java:1124: warning - Tag link: reference not found: Autocomplete.LoginStorageDelegate.onLoginFetch onLoginFetch
```

on my mac, this is the same error:

```
/Users/asferro/workspace/mozilla-central/mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoRuntimeSettings.java:1118: error: reference not found
     * {link Autocomplete.LoginStorageDelegate.onLoginFetch onLoginFetch}.
```

Notice the `-` character after `warning` in the first example and the `:`
character after the second example.

To be able to parse both this patch simiplifies the regex and also adds a
failsafe that will mark anything that contains the string "error" and "warning"
as such.

Differential Revision: https://phabricator.services.mozilla.com/D77545

UltraBlame original commit: eb654059cd8baa99ade5008b2b5bd93166604c44