Bug 1636185: Don't start emitting frames until we can set our initial RID filter. r=jib
This test was flaky because until we set a rid filter, all ssrcs were being
received. This meant that the resolution could have been consistent with any of
foo, bar, or baz when we started looking for resize events. In some cases, that
meant that we would never see the resize event for the transition to bar,
because the video element was already at that resolution.

Depends on D74279

Differential Revision: https://phabricator.services.mozilla.com/D77712

UltraBlame original commit: b38d112a2d6ae41c6039edfca99983216656bc3b