Bug 1642517: Resolves |mach try scriptworker| py3 issue r=rstewart
When python2 interprets headers here, it will attempt to encode values into a string in a blanket fashion.

In this case, the `Accept` header is encoded into an invalid value, but it "works".

Python3 validates more strictly, and throws an error when an array is encountered as a value here. Formatting it
as a valid string instead resolves the problem.

Differential Revision: https://phabricator.services.mozilla.com/D77735

UltraBlame original commit: d701679954b87b4a4ef9c95ed37a5a34bd790cb2