Bug 1642430 - Improve SVGTests implementation r=dholbert
This is should not be a functional change, just making the methods in SVGTests easier to use/understand.
There's a little dead code removal since we can go through the same code path for empty intl.accept_languages as we do when it has a value, and we'd almost always expect it to have a value.

Differential Revision: https://phabricator.services.mozilla.com/D77694

UltraBlame original commit: ec3b786afe744aed08d454a180dfb5a1a70aba5a