Bug 1643264 - Failed ResolveWithTRR requests should not be retried to get the TTL r=dragana,necko-reviewers
While doing the TRR dry run experiment it became apparent that we are issuing
a lot of TRR requests that are causing us to DDOS our TRR provider.
We had a hint that this may only be affecting Windows. Upon inspection of the
code and the APIs we use to measure the performance of each TRR service we
discovered that the codepath that is used to get the TTL for DNS records on
Windows was wrongly exercised for TRR too, causing us to keep repeating a
failing TRR request until success (or until the browser was shut down).

We confirmed this with a unit test and logging.

This patch fixes the condition that caused us to look for failed TRR requests.
We also add a second check that ResolveWithTRRServer calls fail if they also
have the RESOLVE_DISABLE_TRR flag set.

Differential Revision: https://phabricator.services.mozilla.com/D78229

UltraBlame original commit: 2635d8ec9079b2f1721549098e2831dba23e26b1