Bug 1643650 - Round nursery size parameters once when they are set rather than on every use r=sfink
This has the side effect that gcparam() will report the values that are actually in use rather than the value that are set, which seems like a good thing. I had to update test code expectations due to this.

Differential Revision: https://phabricator.services.mozilla.com/D78500

UltraBlame original commit: 76a609e569bfc1dff136361911ea1f667b268b25