Bug 1638711 - Allow test_frameNavigation to use the error event to detect when a load was blocked. r=ckerschb
The current state is that we fire error events for content blocking if the error happens synchronously and src was set when the iframe was in-document, or if the error happens asynchronously (from the parent process).
This test is currently setting src before appending the iframe to the document, and thus was expecting no error event to be fired. We have other content security tests that do rely on the error event being fired.

Since we're doing security checks in the parent, the error event now fires, and this changes the test to report success in that case.

Differential Revision: https://phabricator.services.mozilla.com/D75722

UltraBlame original commit: ca1e2087c0afe59c4c26bf23c6132d6f37d402d8