Bug 1624726: Part 1 - Remove argument type checking in WebGL remoting r=jgilbert
Type checking these queues turned out not to be useful.  It added tokens to the stream that validated the type of the data in the stream against the type of the objects that it would deserialize into.  However, the IPC mechanism is already completely type safe at the source code level so it was not useful in debugging.

Differential Revision: https://phabricator.services.mozilla.com/D78539

UltraBlame original commit: cb87087c351fd01ca9f9e6ab329b0f6b8b398b17