Bug 1644389 - Simplify nsGenericHTMLElement::GetFormControlFrame. r=smaug
This is just cleanup and should have no behavior changes. In the future,
I may optimize GetPrimaryFrame to not flush frames in the whole document
if the style of the node and ancestors haven't changed.

I think one bit of this function shouldn't be needed, but I have left it
as is for now.

Differential Revision: https://phabricator.services.mozilla.com/D78886

UltraBlame original commit: be1c97b1bcf521cd39df0f3cf19704bf90b961fd