Bug 1644264 - Mark frame with text-overflow ellipsis as modified when overflow changes. r=mattwoodrow
This prevents us not descending to rebuild the display list if the
overflow changes during an intermediate layout like a flex layout.

To capture Matrix discussion: This is a bit unfortunate, but the idea is
that text-overflow is usually pretty small (at most one line of content)
so we can probably live with this.

An alternative would be to do something like
nsDisplayListBuilder::MarkFrameForDisplayIfVisible, which could work as
well.

Note that this is a bit of a speculative fix because neither me or
Cameron have been able to reproduce the issue in local builds (I tried
all the combinations of gfx-things that I could think about). :-(

If this doesn't fix it on next nightly, we should back out and try to
repro some other way, I guess. But the hypothesis of why it happens
makes sense to me, and if it's correct it should fix the issue.

Differential Revision: https://phabricator.services.mozilla.com/D79009

UltraBlame original commit: ba429413f7a67002fb50041fd6c3ab37f8d16e2f