Bug 1643869 [wpt PR 24022] - [css-grid] Considering baseline shim when computing max-content size, a=testonly
Automatic update from web-platform-tests
[css-grid] Considering baseline shim when computing max-content size

When computing the MaxContentForChild size for the tracks by the
IndefiniteStrategy we were not considering the baseline shim, since we
assumed it was enough to do it for the MinContentSize.

However, there are possible definitions of the grid track that cause
the MinContentForChild function to not be called, leading to an
incorrect intrinsic size for the grid.

Bug: 1091944
Change-Id: I91a35852dd044ded0fc14f03e97d3e140d80ba12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233239
Reviewed-by: Manuel Rego <regoigalia.com>
Commit-Queue: Javier Fernandez <jfernandezigalia.com>
Cr-Commit-Position: refs/heads/master{#776040}

--

wpt-commits: 13048182561e7b529650fa0b746264c8f289d597
wpt-pr: 24022

UltraBlame original commit: a16cd733e4a0ea1da28922b458053f957800175f