Bug 1644594 [wpt PR 24075] - Add additional Web Platform Tests for the Split Cache, a=testonly
Automatic update from web-platform-tests
Add additional Web Platform Tests for the Split Cache

This expands the existing web platform tests to ensure that same-origin
frames do not share a cache if their top-level origins are different.
The additional tests create a popup with an iframe, where the iframe is
same-origin to the opener. When the popup's mainframe is also
same-origin to the opener, we expect the cache to  be shared. Otherwise,
it should not be.

As with the existing tests, two virtual test suites test the cases of a
split and non-split cache. The default test is disabled as its result
depends on whether the split cache is enabled.

Bug: 1064765
Change-Id: If301e03e0a07a681f960a0972c02fd452b19d798
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238306
Reviewed-by: Shivani Sharma <shivanishachromium.org>
Commit-Queue: Alex Turner <alexmtchromium.org>
Cr-Commit-Position: refs/heads/master{#776663}

--

wpt-commits: 9b572e3991481034103c91d4089b342a934b1a45
wpt-pr: 24075

UltraBlame original commit: 299e8783cb07d23523e2fe0032838a70a4c67a80