Bug 1644343 Part 2 - Make the dimension information in the box model more explicit for screen readers, r=yzen,flod
This patch turns the info that holds the dimensions and position info into a region. Because the visual representation is without explanatory text, an aria-label is constructed for that region to hold a more explicit version of the displayed information.

Differential Revision: https://phabricator.services.mozilla.com/D78952

UltraBlame original commit: d4f83af5e5246671a95ff70c8bbbd67bc2c02fc7