Bug 1644966 - Change second parameter of ConsoleServiceListener to a function. r=jdescottes.
It used to take an object with only 1 property, a function. The way it was
called in the webconsole actor made it harder to follow than it actually need.
We take that as an opportunity to convert the function to a class.

Differential Revision: https://phabricator.services.mozilla.com/D79249

UltraBlame original commit: f48c3a894b8a0a7705500457e3c2831c869c7aea