Bug 1599160 - Wait for less zoom messages in a DAMP test. r=nchevobbe,perftest-reviewers,sparky
DAMP refreshes the page that it loads, and it bypasses the caches, but
we still coalesce in-progress loads, so it's expected to see less of
this.

This actually caught another issue, that I'm fixing in bug 1645180.

Differential Revision: https://phabricator.services.mozilla.com/D79352

UltraBlame original commit: b704a8ee697c2e700dc86d4f698f1b19d25902cd