Bug 1644699 - Allow Debugger.Frame to have GeneratorInfo that is closed. r=arai
The GeneratorInfo data associated with a Debugger.Frame is not guaranteed
to be cleared when a generator transitions from running/suspended to closed.
In the current codebase, we have broadly assumed that to be true, which was
incorrect of us to do.

This patch separates "hasGeneratorInfo" from "isSuspended" so that the two
usecases can be properly separated. This both fixes this bug, and helps make
the codebase clearer about its intent.

Differential Revision: https://phabricator.services.mozilla.com/D79603

UltraBlame original commit: 2d470b425ce9d8fc12c89f087d76ff1a337619ab