Bug 1644536 - Cleanup FieldInitializers initialization. r=mgaudet
FieldInitializers is only defined for functions that have a compiled
enclosing script. We should avoid setting the value when it is not defined.
Update initFromLazyFunction so that fieldInitializers are only read for the
function being delazified. Fix XDR to not track fieldInitializers for lazy
functions with lazy parent functions by checking for exisitance of an
enclosingScope.

Also ensure fieldInitializers are set correctly when cloning scripts. In
practice, we delazify entire script tree before cloning but this is a footgun
none the less.

Differential Revision: https://phabricator.services.mozilla.com/D79282

UltraBlame original commit: cad26877b1f56af9d246fa01e150de91f48be834