Bug 1599769 - Part 2: Update test262 tests. r=jwalden
The previous changeset made a change to test262-update.py. This changeset is
the result of running the script, like this:

    cd js/src/tests
    python test262-update.py --revision 4a8e49b3ca7f9f74a4cafe6621ff9ba548ccc353

The revision is taken from the first line of js/src/tests/test262/GIT-INFO; the
purpose of specifying that revision is to run the script again using the same
input, so that the output reflects only changes to the script itself, not new
tests.

Differential Revision: https://phabricator.services.mozilla.com/D76394

UltraBlame original commit: 0b491b101124323d76ea9d71affeaffd0770d767