Bug 1637632 - Create a compatibility actor and connect it to Inspector's Compatibility panel r=mtigley,daisuke,devtools-backward-compat-reviewers
This patch adds a `CompatibilityActor` that calls the MDNCompatibility
library from the devtool server.
This patch exposes `_declarations` from the `StyleRuleActor` and the
a new function `formDeclaration` responsible for populating the
declarations property.
As a check for correctness, the patch replaces the `pageStyle.getApplied`
call in the Compatibility Panel on the client with `compatibility.getNodeCssIssues`
fetching the array of compatibility issues directly and removing the
dependency on MDNCompatibility library for client side computation of
compatibility issues.

Differential Revision: https://phabricator.services.mozilla.com/D77060

UltraBlame original commit: bbe4cf873819ff7c57828b120f726e0e2607c859