Bug 1645334 - Handle parentProcessTargetActor in WatcherActor. r=jdescottes,ochameau.
In order to handle parentProcessTargetActor in the WatcherActor, we modify
TargetActorRegistry.getTargetActor to return the parent process target when
the passed browsing context id is null.

The comment explaining why we need to call getTargetActor in
WatcherActor#watchResources is modified to better summarize the situation.

We also take this as an opportunity to handle those targets in unwatchResources.

Differential Revision: https://phabricator.services.mozilla.com/D79424

UltraBlame original commit: c124c1c7b2a4df82ff393c15fcb436f1a234d479