Bug 1631545 - Add better tooltip to Reps. r=nchevobbe
This patch adds title attributes on the different Reps, alongside with a new `shouldRenderTooltip` prop.
The props is only used from the ExpressionPanel in the debugger so far.
Tests are added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D72900

UltraBlame original commit: e2438f25cfb36b2cbf5809e47017733d9f3da3dc