Bug 1646633 - Fix intermittent on browser_storage_cache_error.js. r=jdescottes,ochameau.
This patch removes the browsingContextID getter in favor of creating a property
on the instance in the constructor.
The getter was an attempt to handle the fact that we can have multiple browsing
context over time, but it wasn't working that well, so we'll tackle that as
part of Bug 1625027.

Differential Revision: https://phabricator.services.mozilla.com/D80151

UltraBlame original commit: 634266b438b1c14e8f7c94056a6cf945319a0595