Bug 1645997 - Use BrowsingContext to get the top-level window context in PermissionDelegateHandler.cpp r=baku
We used the Document::GetWindowContext() to get the window context
related to the doucment. But, this could return a nullptr if the
document is detached from the window or the docuemnt is destoried.

We move to use the BrowsingContext instead which won't return a nullptr
in these cases.

Differential Revision: https://phabricator.services.mozilla.com/D79897

UltraBlame original commit: 08c6683cd6bec0d7ad40f6fef30460484d7f51e0